pyproject.toml
[project]
name = "foundationmodelsvla"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.7.2",
    "pillow>=12.0.0",
]
